% This is a template R document file
% Distributed under GPL 3 or later

\name{plot.NPsimex}
\alias{plot.NPsimex}
\title{Plot a NPsimex Object}
\description{
  To plot a ``NPsimex'' object generated by simex.density(...), or simexH.density(...).
}
\details{
This function is to plot the estimated function generated by simex.density(...), or simexH.density(...)}
\author{
  X.F. Wang \email{wangx6@ccf.org}
}
\seealso{
 \code{\link{simex.density}, \link{simex.H.density}.
}}
\keyword{smoothing}