% This is a template R document file
% Distributed under GPL 3 or later

\name{print.NPsimex}
\alias{print.NPsimex}
\title{Print a NPsimex Object}
\description{
  To print a ``NPsimex'' object generated by simex.density(...), or simexH.density(...).
}
\details{
This function is to print the summary description from the object generated by simex.density(...), or simexH.density(...)}
\author{
  X.F. Wang \email{wangx6@ccf.org}
}
\seealso{
 \code{\link{simex.density}, \link{simex.H.density}.
}}
\keyword{smoothing}

